gui: Add deletion of scan matches (aka scanmem's `delete`)
Let the user press 'Delete' or 'BackSpace' over the scanresult
treeview to delete the selected match.
Add a right click menu option for the same purpose.

To do this the match id is stored in the scanresult liststore at
update time, then the `delete [id]` command is sent to the
scanmem backend when asked to.